requirements.txt
Pillow~=8.4.0
pip~=21.3.1
wheel~=0.37.0
rsa~=4.7.2
pyasn1~=0.4.8
six~=1.16.0
gast~=0.4.0
numpy~=1.21.4
h5py~=3.6.0
pytz~=2021.3
requests~=2.26.0
keras~=2.7.0
colorama~=0.4.4
tensorflow~=2.7.0
pandas~=1.3.4
wrapt~=1.13.3
setuptools~=59.2.0
tomli~=1.2.2
Django~=3.2.9
sqlparse~=0.4.2
asgiref~=3.4.1
cachetools~=4.2.4
certifi~=2021.10.8
urllib3~=1.26.7
matplotlib~=3.5.0
python-dateutil~=2.8.2
Markdown~=3.3.6
oauthlib~=3.1.1
idna~=3.3
Werkzeug~=2.0.2
fonttools~=4.28.1
pyparsing~=3.0.6
cycler~=0.11.0
packaging~=21.3
kiwisolver~=1.3.2
tensorboard~=2.7.0
flatbuffers~=2.0
astunparse~=1.6.3
termcolor~=1.1.0
django-filter~=21.1
djangorestframework~=3.12.4
DeepImageSearch~=1.4
tqdm~=4.62.3
annoy~=1.17.0
zipp~=3.6.0
djangorestframework-jwt~=1.11.0
PyJWT~=2.3.0